Commit e-mail maturity scanner
The Email Maturity Scanner is made in Visual Studio Code and Deployteq. To have the ability to show all the code that is made to my university, I commit the files needed.
The code has been adapted so that the application can be navigated locally. It is not completely functional, that requires Deployteq or a link to a database.